Add page structure and change layout
Added various buttons to the html which will allow the
user to enter the numbers or operator they want, as well
as an "AC" button that will fully reset the display of the
calculator and a "C" button that will delete the last
item to be entered and shown on the display.
Centered the main div container containing all the
elements for the calculator.