Connect HTML btns to JS and add events.
Linked the display screen as well as all the buttons
on the calculator to JS using querySelector in order
to allow the user to carry out the calculations they want to.
Clicking any of the number buttons will input said number onto
the display screen to let the user know what they are inputting.
AC button will clear the display fully if the user chooses to do so.
C button will remove the character last entered by the user if
they entered something by mistake. Operator buttons will be disabled
if display is currently showing "ERROR!". "ERROR!" can only
be caused by if the user attempts to divide a value with 0.